--- src/main/java/com/example/practice1/Practice1Application.java
@@ -1,12 +1,0 @@
-package com.example.practice1;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Practice1Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Practice1Application.class, args);
-	}
-}
