--- src/main/java/com/example/practice1/auth/controller/AuthController.java
@@ -1,35 +1,0 @@
-package com.example.practice1.auth.controller;
-
-import com.example.practice1.auth.controller.dto.LoginRequest;
-import com.example.practice1.auth.controller.dto.LoginResponse;
-import com.example.practice1.auth.service.AuthService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/auth")
-public class AuthController {
-
-  private final AuthService authService;
-
-  @PostMapping("/login")
-  @ResponseStatus(HttpStatus.OK)
-  public LoginResponse login(@RequestBody LoginRequest request) {
-    return authService.login(request);
-  }
-
-  @PostMapping("/logout")
-  @ResponseStatus(HttpStatus.NO_CONTENT)
-  public void logout(@RequestHeader("Authorization") String token) {
-    authService.logout(token);
-  }
-
-
-}
