--- src/main/java/com/example/practice1/auth/controller/dto/LoginRequest.java
@@ -1,5 +1,0 @@
-package com.example.practice1.auth.controller.dto;
-
-public record LoginRequest(String loginId, String password) {
-  
-}
