--- src/main/java/com/example/practice1/member/controller/dto/MemberRequest.java
@@ -1,6 +1,0 @@
-package com.example.practice1.member.controller.dto;
-
-public record MemberRequest(String loginId, String password, String name) {
-
-
-}
