--- src/main/java/com/example/practice1/post/controller/dto/PostResponse.java
@@ -1,11 +1,0 @@
-package com.example.practice1.post.controller.dto;
-
-import com.example.practice1.post.controller.entity.PostEntity;
-
-public record PostResponse(Long id, String title, String content, String author) {
-
-  public static PostResponse fromEntity(PostEntity postEntity) {
-    return new PostResponse(postEntity.getId(), postEntity.getTitle(), postEntity.getContent(),
-        postEntity.getAuthor());
-  }
-}
