--- src/main/java/com/example/practice1/post/service/HelloService.java
@@ -1,17 +1,0 @@
-package com.example.practice1.post.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class HelloService {
-
-
-  public String getString() {
-    return "hello String";
-  }
-
-  // 이 파일은 서비스다! 라고 알려준다.
-  // cmd + e : 파일 빨리 찾기
-
-
-}
