+++ src/main/java/ssafy/study/ssafystudy/Practice1Application.java
@@ -1,0 +1,12 @@
+package ssafy.study.ssafystudy;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Practice1Application {
+
+  public static void main(String[] args) {
+    SpringApplication.run(Practice1Application.class, args);
+  }
+}
