+++ src/main/java/ssafy/study/ssafystudy/domain/auth/controller/dto/LoginRequest.java
@@ -1,0 +1,5 @@
+package ssafy.study.ssafystudy.domain.auth.controller.dto;
+
+public record LoginRequest(String loginId, String password) {
+
+}
