+++ src/main/java/ssafy/study/ssafystudy/domain/member/controller/dto/MemberRequest.java
@@ -1,0 +1,6 @@
+package ssafy.study.ssafystudy.domain.member.controller.dto;
+
+public record MemberRequest(String loginId, String password, String name) {
+
+
+}
