+++ src/main/java/ssafy/study/ssafystudy/domain/member/controller/dto/MemberResponse.java
@@ -1,0 +1,12 @@
+package ssafy.study.ssafystudy.domain.member.controller.dto;
+
+import ssafy.study.ssafystudy.domain.member.entity.MemberEntity;
+
+public record MemberResponse(Long id, String loginId, String name) {
+
+  public static MemberResponse fromEntity(MemberEntity memberEntity) {
+    return new MemberResponse(memberEntity.getId(),
+        memberEntity.getLoginId(), memberEntity.getName());
+  }
+
+}
