+++ src/main/java/ssafy/study/ssafystudy/global/exception/CustomException.java
@@ -1,0 +1,16 @@
+package ssafy.study.ssafystudy.global.exception;
+
+import lombok.Getter;
+import ssafy.study.ssafystudy.global.exception.error.ErrorCode;
+
+@Getter
+public class CustomException extends RuntimeException {
+
+  private ErrorCode errorCode;
+
+  public CustomException(ErrorCode errorCode) {
+    super(errorCode.getMessage());
+    this.errorCode = errorCode;
+  }
+
+}
