+++ src/main/java/ssafy/study/ssafystudy/global/exception/GlobalExceptionHandler.java
@@ -1,0 +1,26 @@
+package ssafy.study.ssafystudy.global.exception;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import ssafy.study.ssafystudy.global.exception.error.ErrorCode;
+import ssafy.study.ssafystudy.global.response.ApiResponse;
+
+@Slf4j
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+  @ExceptionHandler(CustomException.class)
+  public ResponseEntity<ApiResponse<Void>> handleCustomException(CustomException e) {
+    ErrorCode errorCode = e.getErrorCode();
+    return ResponseEntity.status(errorCode.getHttpStatus()).body(ApiResponse.error(errorCode));
+  }
+
+  @ExceptionHandler(Exception.class)
+  public ResponseEntity<ApiResponse<Void>> handleException(Exception e) {
+    log.error("Unexpected exception occurred", e);
+    return ResponseEntity.internalServerError()
+        .body(ApiResponse.error(ErrorCode.INTERNAL_SERVER_ERROR));
+  }
+}
