--- src/test/java/com/example/practice1/Practice1ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.practice1;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Practice1ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
