+++ src/test/java/ssafy/study/ssafystudy/Practice1ApplicationTests.java
@@ -1,0 +1,13 @@
+package ssafy.study.ssafystudy;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Practice1ApplicationTests {
+
+  @Test
+  void contextLoads() {
+  }
+
+}
